feat(metadata): re-implement metadata
This commit re-implements the metadata library so that it does not
alter objects, and so that it is more independent of AtScript and
TypeScript. As a result, we do not support AtScript out of the box for
it’s annotations, but it can be enabled by calling
addFunctionMetadataLocation(‘annotate’) or
addFunctionMetadataLocation(‘annotations’) depending on which version
of the compiler you are using for your code. A similar api call could
be made for TypeScript support of annotations, if they add it.

For aurelia, those who need to supply metadata should now use a
“metadata” function or property instead of an “annotations” property.