feat(all): new decorator infrastructure
BREAKING CHANGE We do away with the static metadata and instead now
have a static decorators method or property. Developers use the
Decorators helper to set up decorators. These are ES7 spec compliant
decorators that will be applied to the class by the DI system. If your
language supports Decorators, such as ES7 or TypeScript 1.5+, then you
don’t need to use the static method or the helper. Instead, just import
the decorators themselves and use them on your classes, per the
decorator spec. This new capability allows registering any class
decorator with the Decorators helper so that we can enable the use of
class decorators in non-ES7/TypeScript scenarios.